core/requirements-dev.txt
pydantic
pytest
pyyaml
